third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_boundary_events_on_image_map-expected.txt

This is a testharness.js-based test.
[FAIL] pointer boundary events when the <area> is resized
  assert_equals: expected "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerleave@map#map1, pointerover@img#img1, pointerenter@img#img1, pointerout@img#img1, pointerleave@img#img1, pointerleave@div#container]" but got "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerleave@map#map1, pointerleave@div#container]"
[FAIL] pointer boundary events when usemap is modified
  assert_equals: expected "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerleave@map#map1, pointerover@area#area2-1, pointerenter@map#map2, pointerenter@area#area2-1, pointerout@area#area2-1, pointerleave@area#area2-1, pointerleave@map#map2, pointerleave@div#container]" but got "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerleave@map#map1, pointerleave@div#container]"
[FAIL] pointer boundary events when new <area> is available
  assert_equals: expected "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerover@area#area1-2, pointerenter@area#area1-2, pointerout@area#area1-2, pointerleave@area#area1-2, pointerleave@map#map1, pointerleave@div#container]" but got "[pointerover@area#area1-1, pointerenter@div#container, pointerenter@map#map1, pointerenter@area#area1-1, pointermove@area#area1-1, pointerout@area#area1-1, pointerleave@area#area1-1, pointerleave@map#map1, pointerleave@div#container]"
Harness: the test ran to completion.
